Split up ArrayObj
